gxp-mqtt-worker/requirements.txt
paho-mqtt==2.1.0
supabase==2.6.0
urllib3<2.2
python-dotenv==1.0.0
